+++ packages/smarthr-ui/src/components/Textarea/client/Textarea.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen, waitFor } from '@testing-library/react'
 import { useEffect, useState } from 'react'
 
-import { IntlProvider } from '../../intl'
-import { FormControl } from '../FormGroup'
+import { IntlProvider } from '../../../intl'
+import { FormControl } from '../../FormGroup'
 
 import { Textarea } from './Textarea'
