+++ packages/smarthr-ui/src/components/Textarea/client/index.ts
@@ -1,0 +1,1 @@
+export { Textarea } from './Textarea'
